planner: move some ut into new package
