+++ pkg/planner/core/BUILD.bazel
@@ -248,7 +248,6 @@ go_test(
         "plan_test.go",
         "plan_to_pb_test.go",
         "planbuilder_test.go",
-        "point_get_plan_test.go",
         "preprocess_test.go",
         "rule_generate_column_substitute_test.go",
         "rule_join_reorder_dp_test.go",
@@ -275,7 +274,6 @@ go_test(
         "//pkg/infoschema",
         "//pkg/kv",
         "//pkg/meta/model",
-        "//pkg/metrics",
         "//pkg/parser",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
@@ -331,7 +329,6 @@ go_test(
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_tipb//go-tipb",
-        "@com_github_prometheus_client_model//go",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
