+++ pkg/planner/core/tests/pointget/point_get_plan_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package core_test
+package pointget
 
 import (
 	"context"
